requirements.txt
fastapi
uvicorn
python-dotenv
torch
torchaudio
librosa
numpy<2.0.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
transformers
pydub
scipy
speechbrain
huggingface_hub<0.20.0
soundfile
pyyaml
joblib
scikit-learn
